BPR-121: add command for lines of comments per file
